pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vectorbench"
version = "0.1.0"
description = "Embedding Strategy Experimentation Platform"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32",
    "sentence-transformers>=2.7",
    "chromadb>=0.5",
    "plotly>=5.20",
    "umap-learn>=0.5",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pandas>=2.2",
    "joblib>=1.3",
    "nltk>=3.8",
    "torch>=2.2",
    "transformers>=4.40",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.hatch.build.targets.wheel]
packages = ["vectorbench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning:chromadb",
]

[tool.coverage.run]
source = ["vectorbench"]
omit = ["tests/*", "app/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
